* develop:
  Wed Jun 22 01:30:24 WEST 2016
  Revert Displayable.java in commit ccda961ea0a64c2a550a7230adcc9ae8fb7ef75b.
  Fixed EndlessRecyclerOnScrollListener.java bug.
  Fixed Displayable.java bug on small screen sizes.
  Tue Jun 21 01:29:51 WEST 2016
  move cache control flag from constructor to calling method
  implement web service to retrieve user timeline considering articles, apps updates, store latest apps and features
  Mon Jun 20 01:30:02 WEST 2016
  Sun Jun 19 01:30:25 WEST 2016
  Sat Jun 18 01:30:03 WEST 2016
  Fri Jun 17 01:30:18 WEST 2016
  observed scheduler is specified by calling classes instead of networking client, default OkHttpClient creation method moved to OkHttpClientFactory and base web service url exposed as dependency in constructor
  refactor web services to expose http client and converter as dependencies
  Thu Jun 16 01:30:22 WEST 2016

# Conflicts:
#	v8engine/src/main/java/cm/aptoide/pt/v8engine/fragment/implementations/AppViewFragment.java

+++ networkclient/src/main/java/cm/aptoide/pt/networkclient/okhttp/cache/RequestCache.java
@@ -62,13 +62,21 @@ public class RequestCache {
 
 			if( BuildConfig.DEBUG && cachePath.exists() ) {
 				int deletedFiles = 0;
-				for(File f : cachePath.listFiles()) {
-					deletedFiles += f.delete() ? 1 : 0;
+				File[] cacheFiles = cachePath.listFiles();
+				if(cacheFiles!=null && cacheFiles.length>0) {
+					for(File f : cacheFiles) {
+						deletedFiles += f.delete() ? 1 : 0;
+					}
+					deletedFiles += cachePath.delete() ? 1 : 0;
 				}
-				deletedFiles += cachePath.delete() ? 1 : 0;
-				Log.w(TAG, String.format("cache running in debug mode : cleaned %d disk cache " +
-						"files",
-						deletedFiles));
+
+				Log.w(
+						TAG,
+						String.format(
+								"cache running in debug mode : cleaned %d disk cache files",
+								deletedFiles
+						)
+				);
 			}
 
 			diskLruCache = DiskLruCache.open(
